requirements.txt
discord.py[voice]
asyncpraw